Expose using a different API
